requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.27.0,<0.28
geoserver-rest>=2.7.0,<2.8